pyproject.toml
[project]
name = "deck-of-dragons"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0",
    "gdelt>=0.1.10",
    "sentence-transformers>=3.0",
    "scikit-learn>=1.4",
    "spacy>=3.7",
    "transformers>=4.40",
    "torch>=2.2",
    "textual>=0.61",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
